requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# External Services
redis==5.0.1
twilio==8.10.0
httpx==0.25.2

# File Uploads
python-multipart==0.0.6
aiofiles==23.2.1

# Type Checking & Code Quality
mypy==1.7.1
types-sqlalchemy==1.4.53.38
types-bleach>=6.0.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Code Formatting & Linting
black==23.11.0
flake8==6.1.0

# Environment Management
python-dotenv==1.0.0

# Utilities
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.2.0

# Security & Sanitization
bleach>=6.0.0

# Background Tasks
celery>=5.3.0

# ML Dependencies
scikit-learn>=1.3.0
numpy>=1.24.0
scipy>=1.11.0
